change expiration from integer to long

+++ spring-boot-rabbitmq-delay-queue/src/main/java/com/wooyoo/blog/spring_boot/rabbitmq/delay/queue/ExpirationMessagePostProcessor.java
@@ -8,9 +8,9 @@ import org.springframework.amqp.core.MessagePostProcessor;
  * 设置队列消息的过期时间
  */
 public class ExpirationMessagePostProcessor implements MessagePostProcessor {
-    private final Integer ttl; // 毫秒
+    private final Long ttl; // 毫秒
 
-    public ExpirationMessagePostProcessor(Integer ttl) {
+    public ExpirationMessagePostProcessor(Long ttl) {
         this.ttl = ttl;
     }
